Changed to Decimal

+++ BudgetPlanner.Models/Budget/BudgetCreate.cs
@@ -16,7 +16,7 @@ namespace BudgetPlanner.Models
         [Required]
         [Display(Name = "Budget Amount")]
         [DisplayFormat(DataFormatString = "{0:C}")]
-        public double BudgetAmount { get; set; }
+        public decimal BudgetAmount { get; set; }
 
         public virtual List<int> ListOfTransactionIds { get; set; } = new List<int>();
