+++ BudgetPlanner.Models/Budget/BudgetEdit.cs
@@ -18,7 +18,7 @@ namespace BudgetPlanner.Models
 
         [Display(Name = "Budgeted Amount")]
         [DisplayFormat(DataFormatString = "{0:C}")]
-        public double BudgetAmount { get; set; }
+        public decimal BudgetAmount { get; set; }
        
         //public List<int> ListOfCategoryIds { get; set; } = new List<int>();
